requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
attrs==25.1.0
click==8.1.8
colorama==0.4.6
coverage==7.6.12
dynaconf==3.2.10
fastjsonschema==2.21.1
importlib_metadata==8.6.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
nbformat==5.10.4
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-magic==0.4.27
python-magic-bin==0.4.14  # For Windows compatibility
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
pywin32==308
referencing==0.36.2
rpds-py==0.23.1
setuptools==75.8.0
structlog==25.1.0
tenacity==9.0.0
traitlets==5.14.3
typing_extensions==4.12.2
zipp==3.21.0

# Document parsing dependencies
pypdf>=3.17.1
python-docx>=0.8.11
odfpy>=1.4.1
EbookLib>=0.17.1
aiofiles>=0.8.0
python-magic>=0.4.27

# Spreadsheet parsing dependencies
pandas>=2.0.0
openpyxl>=3.1.0
odfpy>=1.4.1

# Code and configuration parsing dependencies
tomli>=2.0.1  # For TOML parsing
PyYAML>=6.0.1  # For YAML parsing
xmltodict>=0.13.0  # For XML parsing

# Vector and CAD parsing dependencies
ezdxf>=1.0.0
svglib>=1.5.1
reportlab>=4.0.4  # Required by svglib

# Archive handling dependencies
py7zr>=0.20.5
rarfile>=4.0
python-magic>=0.4.27


Kreuzberg